full breakout of classes
